% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servcatr.R
\name{sc_profiles}
\alias{sc_profiles}
\title{Query ServCat reference profiles}
\usage{
sc_profiles(references)
}
\arguments{
\item{references}{a vector of numeric reference codes for ServCat records}
}
\value{
a list containing the contents of the parsed ServCat ari request and the api response
}
\description{
Query ServCat reference profiles
}
\examples{
\dontrun{
sc_profiles(87544)
sc_profiles(c(87544, 87545))
}
}
